+++ src/Internal/Ast/ContractVisitor.php
@@ -2,15 +2,11 @@
 
 declare(strict_types=1);
 
-namespace TypePHP\Internal;
+namespace TypePHP\Internal\Ast;
 
 use PhpParser\Node;
 use PhpParser\NodeVisitorAbstract;
 use TypePHP\Contract\DocblockExtractor;
-use TypePHP\Internal\Visitor\FunctionContractInjector;
-use TypePHP\Internal\Visitor\NodeBuilder;
-use TypePHP\Internal\Visitor\PropertyHookInjector;
-use TypePHP\Internal\Visitor\ScopeManager;
 
 /**
  * @internal AST Node Visitor that injects contract checks, scope tracking, property hook validation, and parameter/return wrappers into functions and methods.
