+++ src/Internal/Ast/ScopeCleaner.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TypePHP\Internal;
+namespace TypePHP\Internal\Ast;
 
 use TypePHP\Resolver\TemplateManager;
